requirements.txt
pip
matplotlib~=3.5.0
tensorflow-gpu
scikit-learn
scikit-image~=0.18.3
numpy~=1.21.4
Pillow~=9.0.0